feat(appeals service api tests): split out intregration and unit tests
Split out integration and unit tests on the appeals service api so integration tests can be run only
on commits to main

AS-5961
