+++ package.json
@@ -45,7 +45,9 @@
 		"test:common": "cd packages/common; npm run test",
 		"test:documents-api": "cd packages/document-service-api; npm run test",
 		"test:pdf-api": "cd packages/pdf-service-api; npm run test",
-		"test:web": "cd packages/forms-web-app; npm run test"
+		"test:web": "cd packages/forms-web-app; npm run test",
+		"integration-test": "run-s --continue-on-error --print-label \"integration-test:*\"",
+		"integration-test:appeals-api": "cd packages/appeals-service-api; npm run integration-test"
 	},
 	"repository": {
 		"type": "git",
