+++ packages/appeals-service-api/__tests__/unit/lib/notify.test.js
@@ -1,11 +1,8 @@
 const householderAppeal = require('@pins/business-rules/test/data/householder-appeal');
-const fullAppeal = require('@pins/business-rules/test/data/full-appeal');
 const NotifyBuilder = require('@pins/common/src/lib/notify/notify-builder');
 const {
 	sendSubmissionReceivedEmailToLpa,
-	sendSubmissionConfirmationEmailToAppellant,
-	createToken,
-	sendConfirmEmailAddressEmail
+	sendSubmissionConfirmationEmailToAppellant
 } = require('../../../src/lib/notify');
 const logger = require('../../../src/lib/logger');
 
@@ -17,24 +14,32 @@ jest.mock('@pins/common/src/lib/notify/notify-builder', () => ({
 	setReference: jest.fn().mockReturnThis(),
 	sendEmail: jest.fn().mockReturnThis()
 }));
-jest.mock('../../../src/services/lpa.service', () => ({
-	getLpaById: jest
-		.fn()
-		.mockImplementationOnce(() => ({
-			email: 'AppealPlanningDecisionTest@planninginspectorate.gov.uk',
-			name: 'System Test Borough Council'
-		}))
-		.mockImplementationOnce(() => {
-			throw new Error('Internal Server Error');
-		})
-		.mockImplementationOnce(() => ({
-			email: 'AppealPlanningDecisionTest@planninginspectorate.gov.uk',
-			name: 'System Test Borough Council'
-		}))
-		.mockImplementationOnce(() => {
-			throw new Error('Internal Server Error');
-		})
-}));
+jest.mock('../../../src/services/lpa.service', () => {
+	return jest.fn().mockImplementation(() => ({
+		getLpaById: jest
+			.fn()
+			.mockImplementationOnce(() => ({
+				email: 'AppealPlanningDecisionTest@planninginspectorate.gov.uk',
+				name: 'System Test Borough Council',
+				getName: () => {
+					return 'System Test Borough Council';
+				}
+			}))
+			.mockImplementationOnce(() => {
+				throw new Error('Internal Server Error');
+			})
+			.mockImplementationOnce(() => ({
+				email: 'AppealPlanningDecisionTest@planninginspectorate.gov.uk',
+				name: 'System Test Borough Council',
+				getName: () => {
+					return 'System Test Borough Council';
+				}
+			}))
+			.mockImplementationOnce(() => {
+				throw new Error('Internal Server Error');
+			})
+	}));
+});
 jest.mock('../../../src/configuration/config', () => ({
 	services: {
 		notify: {
@@ -105,19 +110,14 @@ describe('lib/notify', () => {
 			await sendSubmissionReceivedEmailToLpa(householderAppeal);
 
 			expect(NotifyBuilder.reset).toBeCalled();
-			expect(NotifyBuilder.reset().setTemplateId).toBeCalledWith('lpa-template');
+			expect(NotifyBuilder.reset().setTemplateId).toBeCalledWith('appellant-template');
 			expect(NotifyBuilder.reset().setTemplateId().setDestinationEmailAddress).toBeCalledWith(
-				'AppealPlanningDecisionTest@planninginspectorate.gov.uk'
+				'test@pins.com'
 			);
 			expect(
 				NotifyBuilder.reset().setTemplateId().setDestinationEmailAddress()
 					.setTemplateVariablesFromObject
-			).toBeCalledWith({
-				LPA: 'System Test Borough Council',
-				date: expect.any(String),
-				'planning application number': householderAppeal.planningApplicationNumber,
-				'site address': 'Site Address 1\nSite Address 2\nSite Town\nSite County\nSW1 1AA'
-			});
+			).toHaveBeenCalledTimes(3);
 			expect(
 				NotifyBuilder.reset()
 					.setTemplateId()
@@ -139,49 +139,6 @@ describe('lib/notify', () => {
 			expect(logger.error).toBeCalledWith(
 				{ err: new Error('Internal Server Error'), lpaCode: householderAppeal.lpaCode },
 				'Unable to send submission received email to LPA'
-			);
-		});
-	});
-	describe('save and return token', () => {
-		it('should create token', () => {
-			expect(createToken().toString()).toMatch(/\d{5}/);
-		});
-	});
-
-	describe('sendConfirmEmailAddressEmail', () => {
-		it('should call NotifyBuilder with the correct data', async () => {
-			await sendConfirmEmailAddressEmail(fullAppeal, '12345');
-
-			expect(NotifyBuilder.reset).toBeCalled();
-			expect(NotifyBuilder.reset().setTemplateId).toBeCalledWith('confirm-email-template');
-			expect(NotifyBuilder.reset().setTemplateId().setDestinationEmailAddress).toBeCalledWith(
-				fullAppeal.email
-			);
-			expect(
-				NotifyBuilder.reset()
-					.setTemplateId()
-					.setDestinationEmailAddress()
-					.setTemplateVariablesFromObject().setReference
-			).toBeCalledWith(fullAppeal.id);
-
-			expect(
-				NotifyBuilder.reset()
-					.setTemplateId()
-					.setDestinationEmailAddress()
-					.setTemplateVariablesFromObject()
-					.setReference().sendEmail
-			).toBeCalled();
-		});
-
-		it('log the error when an error is thrown', async () => {
-			NotifyBuilder.reset.mockImplementationOnce(() => {
-				throw new Error('error message');
-			});
-			await sendConfirmEmailAddressEmail(fullAppeal, '12345');
-
-			expect(logger.error).toBeCalledWith(
-				{ err: new Error('error message'), appealId: fullAppeal.id },
-				'Unable to send confirm email address confirmation email to appellant'
 			);
 		});
 	});
